Use UTF-8 for .ts files
With QT5 the default encoding changed from Latin1 to UTF-8 for .ts
files. Additionally the only supported encodings by the localization
tools are UTF-8 and UTF-16 now.
As a result when using QT5's lupdate on the project several strings
cannot be matched any more, resulting in an incompletely translated GUI.
Change the translation files to UTF-8 and add the appropriate definition
to the project file for backwards compatibility with QT4.